refactor: switch to recursive project structure
