+++ src/components/Tasks/components/TaskList/TaskList.tsx
@@ -1,6 +1,6 @@
 import React, { FC } from 'react'
 
-import { Task } from '../Task'
+import { Task } from './components/Task'
 
 import { TaskListProps } from 'types'
