+++ src/components/shared/SVGIcon/index.tsx
